+++ forest/fadewidget.h
@@ -30,7 +30,7 @@
 #include <QDesktopWidget>
 #include <QSettings>
 
-#include "../library/miscutills/miscutills.h"
+#include "miscutills.h"
 
 
 class fadewidget : public QWidget